+++ stream_membership/_typing.py
@@ -1,0 +1,1 @@
+CoordinateName = str | tuple[str, str]
